+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## [0.3.1]
 
+- using python executable to install packages instead of pip
 - added `/map` and `/map-refresh` commands for repository mapping functionality
 - prevent infinite loading state after application refresh